MM-10111: Omits PostBodyAdditionalContent if post is deleted.
