+++ components/rhs_root_post/rhs_root_post.jsx
@@ -18,15 +18,14 @@ import * as Utils from 'utils/utils.jsx';
 import DotMenu from 'components/dot_menu';
 import EmojiPickerOverlay from 'components/emoji_picker/emoji_picker_overlay.jsx';
 import FileAttachmentListContainer from 'components/file_attachment_list';
-import PostBodyAdditionalContent from 'components/post_view/post_body_additional_content';
 import PostFlagIcon from 'components/post_view/post_flag_icon.jsx';
-import PostMessageContainer from 'components/post_view/post_message_view';
 import ReactionListContainer from 'components/post_view/reaction_list';
 import PostTime from 'components/post_view/post_time.jsx';
 import ProfilePicture from 'components/profile_picture.jsx';
 import EmojiIcon from 'components/svg/emoji_icon';
 import MattermostLogo from 'components/svg/mattermost_logo';
 import ChannelPermissionGate from 'components/permissions_gates/channel_permission_gate';
+import MessageWithAdditionalContent from 'components/message_with_additional_content';
 
 import UserProfile from 'components/user_profile.jsx';
 
@@ -431,18 +430,12 @@ export default class RhsRootPost extends React.Component {
                         </div>
                         <div className='post__body'>
                             <div className={postClass}>
-                                <PostBodyAdditionalContent
+                                <MessageWithAdditionalContent
                                     post={post}
                                     previewCollapsed={this.props.previewCollapsed}
                                     previewEnabled={this.props.previewEnabled}
                                     isEmbedVisible={this.props.isEmbedVisible}
-                                >
-                                    <PostMessageContainer
-                                        post={post}
-                                        isRHS={true}
-                                        hasMention={true}
-                                    />
-                                </PostBodyAdditionalContent>
+                                />
                             </div>
                             {fileAttachment}
                             <ReactionListContainer
